optimize generator

+++ generator/function/generate_single_migrations.php
@@ -1,21 +1,24 @@
 <?php
 
-function generateMigrations ($list, $outputDir = '') {
+function generateSingleMigrations ($list, $outputDir = '') {
+
+$outputDir = $outputDir."/Migrations/";
+if (!file_exists($outputDir)) mkdir($outputDir, 0777, true); // generate folder module
+
+foreach($list as $items){
+    $itemx[] = $items;
+    $moduleSelector =  strtolower(splitUppercaseToUnderscore($items->name));
+
+    // print_r($items); die();
 
 $scriptMigrations = '';
 $reverseList = '';
 $schema  = '';
-foreach($list as $item){
+foreach($itemx as $item){
 
-    // print_r($item); echo "<hr>";
-    logInfoGenerate($item);
     if(!isset($item->name))  die();
  
 	$selector =  strtolower(splitUppercaseToUnderscore($item->name));
-
-    if (isset($item->loging_user) && $item->loging_user == true) {
-        $item->column = enableLogingUser($item->column);
-    }
 
     $columList = '';
     $no = 1;
@@ -71,13 +74,16 @@ foreach($list as $item){
             $table->softDeletes("deleted_at");';
 
  
-    $reverseList .= '        Schema::dropIfExists("'.$selector.'"); ' ."\r\n";
 
-    $schema .= '        Schema::create("'.$selector.'", function (Blueprint $table) {' ."\r\n";
-    $schema .= '        '. $columList .'' ."\r\n";
-    $schema .= ''. $loging_date .'' ."\r\n";
-    $schema .= '        });' ."\r\n\r\n";
-    $schema .= '        '. $fulltext .'' ."\r\n";
+    if ($selector == $moduleSelector) {
+        $reverseList .= '    Schema::dropIfExists("'.$selector.'"); ' ."\r\n";
+
+        $schema .= 'Schema::create("'.$selector.'", function (Blueprint $table) {' ."\r\n";
+        $schema .= '        '. $columList .'' ."\r\n";
+        $schema .= ''. $loging_date .'' ."\r\n";
+        $schema .= '        });' ."\r\n\r\n";
+        $schema .= '        '. $fulltext .'' ."\r\n";
+    }
 
     $scriptMigrations = '<?php
 
@@ -85,7 +91,7 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class FirstInit extends Migration
+class '. $item->name .' extends Migration
 {
     /**
      * Run the migrations.
@@ -114,26 +120,14 @@ class FirstInit extends Migration
 
 $scriptMigrations = str_replace('"',"'", $scriptMigrations);
 // $fname = gmdate("Y_m_d_His", time() + 60 * 60 * 7) . "_update_" . gmdate("YmdHis", time() + 60 * 60 * 7);
-$fname = '2020_11_04_184938_first_init';
-$createMigrations = fopen($outputDir."$fname.php", "w") or die("Unable to open file!");
-fwrite($createMigrations, $scriptMigrations);
-fclose($createMigrations);   
+$fname = gmdate("Y_m_d_His", time() + 60 * 60 * 7) .'_'.str_replace(' ',"_", $moduleSelector);
+$fname = strtolower($fname);
+
+    $createMigrations = fopen($outputDir."$fname.php", "w") or die("Unable to open file!");
+    fwrite($createMigrations, $scriptMigrations);
+    fclose($createMigrations);
 
 }
 
-function logInfoGenerate ($item) {
-    echo "<b style='font-size:23px;' >".$item->name . "</b>  <small style='padding:2px; color:#fff; background:green' > &nbsp; Generated &nbsp; </small><br>";
-    $code = '';
-    foreach ($item->column as $key => $col) {
-        $attr = '';
-        if (isset($col->attributes)) $attr = "".json_encode($col->attributes)."";
-        $code .= "<tr>
-        <td>".$col->name . "</td> 
-        <td>". $col->type."</td> 
-        <td>".$attr."</td>
-        </tr>";
-    }
-    echo "<div style='margin:5px; background:#ccc; padding:5px; border-radius:5px;'><table><tbody>$code</tbody></table></div>";
-    echo "<hr>";
-
+ 
 }
